Zlevels rework

+++ code/modules/admin/verbs/randomverbs.dm
@@ -548,7 +548,12 @@ Traitors and the like can also be revived with the previous role mostly intact.
 		log_admin("[key_name(usr)] deleted [O] at ([O.x],[O.y],[O.z])")
 		message_admins("[key_name_admin(usr)] deleted [O] at ([O.x],[O.y],[O.z])", 1)
 		feedback_add_details("admin_verb","DEL") //If you are copy-pasting this, ensure the 2nd parameter is unique to the new proc!
-		qdel(O)
+
+		//use change turf in case a non-space turf needs to be spawned upon deletion
+		if(isturf(O))
+			O:ChangeTurf(get_base_turf(O.z))
+		else
+			qdel(O)
 
 /client/proc/cmd_admin_list_open_jobs()
 	set category = "Admin"
